Refactor beans_post_meta integration tests to WPCS
